Added powerup store
Added a store which contains powerups for the player to purchase with the coins they have collected. The game pauses when the score is opened. Store can be accessed by clicking the store button on the screen. Store can be closed by clicking this button again. For now, only one powerup exists.