Added high score screen
Top 3 high scores are now tracked and accessible from the main menu screen. Scores are saved in an xml file using Preferences, located in %UserProfile%/.prefs/Coin Getter Preferences on Windows. On Android, SharedPreferences is used and high scores will be deleted when the app is uninstalled but should be kept if the app is updated.