Fixed purchase issue for slower enemy respawn
The conditions for the slower enemy respawn was checking to see if the player had at least 10 coins instead of 100 but still subtracting 100 from the coin count, resulting in a negative coin count. Fixed to check for 100 coins and subtract 100. Other small changes included as well to properly dispose of the pause menu and reformated some stuff in the desktop launcher class.